+++ app/models/rolepermission.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, BigInteger, ForeignKey
+from sqlalchemy import Column, Integer, ForeignKey
 from sqlalchemy.orm import relationship
 from app.database import Base
 
@@ -7,11 +7,11 @@ class RolePermission(Base):
 
     __tablename__ = "role_permissions"
 
-    id = Column(BigInteger, primary_key=True)
+    id = Column(Integer, primary_key=True)
 
-    role_id = Column(BigInteger, ForeignKey("roles.id"))
+    role_id = Column(Integer, ForeignKey("roles.id"))
 
-    permission_id = Column(BigInteger, ForeignKey("permissions.id"))
+    permission_id = Column(Integer, ForeignKey("permissions.id"))
 
     role = relationship("Role", back_populates="permissions")
