+++ app/models/userrole.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, BigInteger, ForeignKey, DateTime
+from sqlalchemy import Column, Integer, ForeignKey, DateTime
 from sqlalchemy.orm import relationship
 from datetime import datetime
 from app.database import Base
@@ -8,11 +8,11 @@ class UserRole(Base):
 
     __tablename__ = "user_roles"
 
-    id = Column(BigInteger, primary_key=True)
+    id = Column(Integer, primary_key=True)
 
-    user_id = Column(BigInteger, ForeignKey("users.id"))
+    user_id = Column(Integer, ForeignKey("users.id"))
 
-    role_id = Column(BigInteger, ForeignKey("roles.id"))
+    role_id = Column(Integer, ForeignKey("roles.id"))
 
     assigned_at = Column(DateTime, default=datetime.utcnow)
